WordPress and plugins update

+++ wp-content/plugins/polylang/settings/settings-module.php
@@ -1,78 +1,151 @@
 <?php
+/**
+ * @package Polylang
+ */
 
 /**
- * base class for all settings
+ * Base class for all settings
  *
  * @since 1.8
  */
 class PLL_Settings_Module {
-	public $active_option, $configure;
-	public $module, $title, $description;
+	/**
+	 * Stores the plugin options.
+	 *
+	 * @var array
+	 */
 	public $options;
-	protected $action_links, $buttons;
 
 	/**
-	 * constructor
+	 * @var PLL_Model
+	 */
+	public $model;
+
+	/**
+	 * Instance of a child class of PLL_Links_Model.
+	 *
+	 * @var PLL_Links_Model
+	 */
+	public $links_model;
+
+	/**
+	 * Stores if the module is active.
+	 *
+	 * @var bool
+	 */
+	public $active_option;
+
+	/**
+	 * Stores the display order priority.
+	 *
+	 * @var int
+	 */
+	public $priority = 100;
+
+	/**
+	 * Stores the module name.
+	 * It must be unique.
+	 *
+	 * @var string
+	 */
+	public $module;
+
+	/**
+	 * Stores the module title.
+	 *
+	 * @var string
+	 */
+	public $title;
+
+	/**
+	 * Stores the module description.
+	 *
+	 * @var string
+	 */
+	public $description;
+
+	/**
+	 * Stores the settings actions.
+	 *
+	 * @var array
+	 */
+	protected $action_links;
+
+	/**
+	 * Stores html fragment for the buttons.
+	 *
+	 * @var array
+	 */
+	protected $buttons;
+
+	/**
+	 * Stores html form when provided by a child class.
+	 *
+	 * @var bool|string
+	 */
+	protected $form = false;
+
+	/**
+	 * Constructor
 	 *
 	 * @since 1.8
 	 *
-	 * @param object $polylang polylang object
-	 * @param array $args
+	 * @param object $polylang Polylang object
+	 * @param array  $args
 	 */
 	public function __construct( &$polylang, $args ) {
 		$this->options = &$polylang->options;
 		$this->model = &$polylang->model;
 		$this->links_model = &$polylang->links_model;
 
-		$args = wp_parse_args( $args, array(
-			'title' => '',
-			'description' => '',
-			'active_option' => false,
-		) );
+		$args = wp_parse_args(
+			$args,
+			array(
+				'title'         => '',
+				'description'   => '',
+				'active_option' => false,
+			)
+		);
 
 		foreach ( $args as $prop => $value ) {
 			$this->$prop = $value;
 		}
 
-		// all possible action links, even if not always a link ;- )
+		// All possible action links, even if not always a link ;-)
 		$this->action_links = array(
-			'configure' => sprintf(
+			'configure'   => sprintf(
 				'<a title="%s" href="%s">%s</a>',
-				__( 'Configure this module', 'polylang' ),
+				esc_attr__( 'Configure this module', 'polylang' ),
 				'#',
-				__( 'Settings', 'polylang' )
+				esc_html__( 'Settings', 'polylang' )
 			),
-
-			'deactivate' => sprintf(
+			'deactivate'  => sprintf(
 				'<a title="%s" href="%s">%s</a>',
-				__( 'Deactivate this module', 'polylang' ),
-				wp_nonce_url( '?page=mlang&amp;tab=modules&amp;pll_action=deactivate&amp;noheader=true&amp;module=' . $this->module, 'pll_deactivate' ),
-				__( 'Deactivate', 'polylang' )
+				esc_attr__( 'Deactivate this module', 'polylang' ),
+				esc_url( wp_nonce_url( '?page=mlang&tab=modules&pll_action=deactivate&noheader=true&module=' . $this->module, 'pll_deactivate' ) ),
+				esc_html__( 'Deactivate', 'polylang' )
 			),
-
-			'activate' => sprintf(
+			'activate'    => sprintf(
 				'<a title="%s" href="%s">%s</a>',
-				__( 'Activate this module', 'polylang' ),
-				wp_nonce_url( '?page=mlang&amp;tab=modules&amp;pll_action=activate&amp;noheader=true&amp;module=' . $this->module, 'pll_activate' ),
-				__( 'Activate', 'polylang' )
+				esc_attr__( 'Activate this module', 'polylang' ),
+				esc_url( wp_nonce_url( '?page=mlang&tab=modules&pll_action=activate&noheader=true&module=' . $this->module, 'pll_activate' ) ),
+				esc_html__( 'Activate', 'polylang' )
 			),
-
-			'activated' => __( 'Activated', 'polylang' ),
-
-			'deactivated' => __( 'Deactivated', 'polylang' ),
+			'activated'   => esc_html__( 'Activated', 'polylang' ),
+			'deactivated' => esc_html__( 'Deactivated', 'polylang' ),
 		);
 
 		$this->buttons = array(
-			'cancel' => sprintf( '<button type="button" class="button button-secondary cancel">%s</button>', __( 'Cancel' ) ),
-			'save'   => sprintf( '<button type="button" class="button button-primary save">%s</button>', __( 'Save Changes' ) ),
+			'cancel' => sprintf( '<button type="button" class="button button-secondary cancel">%s</button>', esc_html__( 'Cancel', 'polylang' ) ),
+			'save'   => sprintf( '<button type="button" class="button button-primary save">%s</button>', esc_html__( 'Save Changes', 'polylang' ) ),
 		);
 
-		// ajax action to save options
-		add_action( 'wp_ajax_pll_save_options', array( &$this, 'save_options' ) );
+		// Ajax action to save options
+		add_action( 'wp_ajax_pll_save_options', array( $this, 'save_options' ) );
 	}
 
 	/**
-	 * tells if the module is active
+	 * Tells if the module is active
 	 *
 	 * @since 1.8
 	 *
@@ -83,9 +156,11 @@ class PLL_Settings_Module {
 	}
 
 	/**
-	 * activates the module
+	 * Activates the module
 	 *
 	 * @since 1.8
+	 *
+	 * @return void
 	 */
 	public function activate() {
 		if ( ! empty( $this->active_option ) ) {
@@ -95,9 +170,11 @@ class PLL_Settings_Module {
 	}
 
 	/**
-	 * deactivates the module
+	 * Deactivates the module
 	 *
 	 * @since 1.8
+	 *
+	 * @return void
 	 */
 	public function deactivate() {
 		if ( ! empty( $this->active_option ) ) {
@@ -107,51 +184,52 @@ class PLL_Settings_Module {
 	}
 
 	/**
-	 * protected method to display a configuration form
+	 * Protected method to display a configuration form
 	 *
 	 * @since 1.8
 	 *
+	 * @return void
 	 */
 	protected function form() {
-		// child classes can provide a form
+		// Child classes can provide a form
 	}
 
 	/**
-	 * public method returning the form if any
+	 * Public method returning the form if any
 	 *
 	 * @since 1.8
 	 *
 	 * @return string
 	 */
 	public function get_form() {
-		static $form = false;
-
-		// read the form only once
-		if ( false === $form ) {
+		// Read the form only once
+		if ( false === $this->form ) {
 			ob_start();
 			$this->form();
-			$form = ob_get_clean();
+			$this->form = ob_get_clean();
 		}
 
-		return $form;
+		return $this->form;
 	}
 
 	/**
-	 * allows child classes to validate their options before saving
+	 * Allows child classes to validate their options before saving
 	 *
 	 * @since 1.8
 	 *
-	 * @param array $options raw options
-	 * @param array options
+	 * @param array $options Raw options
+	 * @return array Options
 	 */
-	protected function update( $options ) {
-		return array(); // it's responsibility of the child class to decide what is saved
+	protected function update( $options ) { // phpcs:ignore VariableAnalysis.CodeAnalysis.VariableAnalysis.UnusedVariable
+		return array(); // It's responsibility of the child class to decide what is saved
 	}
 
 	/**
-	 * ajax method to save the options
+	 * Ajax method to save the options
 	 *
 	 * @since 1.8
+	 *
+	 * @return void
 	 */
 	public function save_options() {
 		check_ajax_referer( 'pll_options', '_pll_nonce' );
@@ -159,30 +237,30 @@ class PLL_Settings_Module {
 			wp_die( -1 );
 		}
 
-		if ( $this->module == $_POST['module'] ) {
-			// it's up to the child class to decide which options are saved, whether there are errors or not
+		if ( isset( $_POST['module'] ) && $this->module === $_POST['module'] ) {
+			// It's up to the child class to decide which options are saved, whether there are errors or not
 			$post = array_diff_key( $_POST, array_flip( array( 'action', 'module', 'pll_ajax_backend', '_pll_nonce' ) ) );
 			$options = $this->update( $post );
 			$this->options = array_merge( $this->options, $options );
 			update_option( 'polylang', $this->options );
 
-			// refresh language cache in case home urls have been modified
+			// Refresh language cache in case home urls have been modified
 			$this->model->clean_languages_cache();
 
-			// refresh rewrite rules in case rewrite,  hide_default, post types or taxonomies options have been modified
-			// don't use flush_rewrite_rules as we don't have the right links model and permastruct
+			// Refresh rewrite rules in case rewrite,  hide_default, post types or taxonomies options have been modified
+			// Don't use flush_rewrite_rules as we don't have the right links model and permastruct
 			delete_option( 'rewrite_rules' );
 
 			ob_start();
 
 			if ( ! get_settings_errors() ) {
-				// send update message
-				add_settings_error( 'general', 'settings_updated', __( 'Settings saved.' ), 'updated' );
+				// Send update message
+				add_settings_error( 'general', 'settings_updated', __( 'Settings saved.', 'polylang' ), 'updated' );
 				settings_errors();
 				$x = new WP_Ajax_Response( array( 'what' => 'success', 'data' => ob_get_clean() ) );
 				$x->send();
 			} else {
-				// send error messages
+				// Send error messages
 				settings_errors();
 				$x = new WP_Ajax_Response( array( 'what' => 'error', 'data' => ob_get_clean() ) );
 				$x->send();
@@ -191,13 +269,15 @@ class PLL_Settings_Module {
 	}
 
 	/**
-	 * get the row actions
+	 * Get the row actions.
 	 *
 	 * @since 1.8
 	 *
-	 * @return array
+	 * @return string[]
 	 */
 	protected function get_actions() {
+		$actions = array();
+
 		if ( $this->is_active() && $this->get_form() ) {
 			$actions[] = 'configure';
 		}
@@ -214,18 +294,18 @@ class PLL_Settings_Module {
 	}
 
 	/**
-	 * get the actions links
+	 * Get the actions links.
 	 *
 	 * @since 1.8
 	 *
-	 * @return array
+	 * @return string[] Action links.
 	 */
 	public function get_action_links() {
 		return array_intersect_key( $this->action_links, array_flip( $this->get_actions() ) );
 	}
 
 	/**
-	 * default upgrade message ( to pro version )
+	 * Default upgrade message ( to Pro version )
 	 *
 	 * @since 1.9
 	 *
@@ -234,14 +314,14 @@ class PLL_Settings_Module {
 	protected function default_upgrade_message() {
 		return sprintf(
 			'%s <a href="%s">%s</a>',
-			__( 'You need Polylang Pro to enable this feature.', 'polylang' ),
+			__( 'To enable this feature, you need Polylang Pro.', 'polylang' ),
 			'https://polylang.pro',
 			__( 'Upgrade now.', 'polylang' )
 		);
 	}
 
 	/**
-	 * allows child classes to display an upgrade message
+	 * Allows child classes to display an upgrade message
 	 *
 	 * @since 1.9
 	 *
@@ -252,11 +332,11 @@ class PLL_Settings_Module {
 	}
 
 	/**
-	 * get the buttons
+	 * Get the buttons.
 	 *
 	 * @since 1.9
 	 *
-	 * @return array
+	 * @return string[] An array of html fragment for the buttons.
 	 */
 	public function get_buttons() {
 		return $this->buttons;
